% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reader.R
\name{docterm_csv}
\alias{docterm_csv}
\title{Create docid term table from directory containing csv files}
\usage{
docterm_csv(fp, fpattern = ".csv")
}
\description{
Create docid term table from directory containing csv files
}
