% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{items}
\alias{items}
\title{Meta data from the Publications of the Institutional Repository of the Max Planck Society}
\format{An object of class \code{"list"}}
\source{
\url{https://pure.mpg.de/rest}
}
\usage{
data(items)
}
\description{
A named list with nodes (publications, sources, persons, organizational units) and edges between them.
}
\keyword{datasets}
