% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{mpis}
\alias{mpis}
\title{Meta data of Max Planck Institutes}
\format{An object of class \code{"list"}}
\source{
\url{https://mpg.de}
}
\usage{
data(mpis)
}
\description{
A named list with nodes (contexts, organizational units, categories, tags) and edges between them.
}
\keyword{datasets}
