% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reader.R
\name{plain_to_corpus}
\alias{plain_to_corpus}
\title{Create corpus from directory containing plain text files}
\usage{
plain_to_corpus(fp, fpattern = ".txt")
}
\description{
Create corpus from directory containing plain text files
}
