% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reader.R
\name{plain_to_dtm}
\alias{plain_to_dtm}
\title{Create document term matrix from directory containing plain text files}
\usage{
plain_to_dtm(fp, fpattern = ".txt")
}
\description{
Create document term matrix from directory containing plain text files
}
