% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{pure}
\alias{pure}
\title{Meta data from the Institutional Repository of the Max Planck Society}
\format{An object of class \code{"list"}}
\source{
\url{https://pure.mpg.de/rest}
}
\usage{
data(pure)
}
\description{
A named list with nodes (contexts, organizational units) and edges between them.
}
\keyword{datasets}
