% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{pure_items}
\alias{pure_items}
\title{Publications from the Institutional Repository of the Max Planck Society}
\format{An object of class \code{"tbl_df"}}
\source{
\url{https://pure.mpg.de/rest}
}
\usage{
data(pure_items)
}
\description{
A data.frame (tibble) with more than 390.000 publications.
}
\keyword{datasets}
