% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{sel}
\alias{sel}
\title{Meta data of Max Planck Institutes}
\format{An object of class \code{"list"}}
\source{
\url{https://pure.mpg.de/rest}
}
\usage{
data(sel)
}
\description{
A named list with nodes (contexts, organizational units) and edges between them.
}
\keyword{datasets}
