% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{sel_items_src}
\alias{sel_items_src}
\title{Sources of Publications from Max Planck Institutes}
\format{An object of class \code{"tbl_df"}}
\source{
\url{https://pure.mpg.de/rest}
}
\usage{
data(sel_items_src)
}
\description{
A data.frame (tibble) with more than 200.000 publication sources (JOURNALS, COLLECTED_EDITIONS, ...).
}
\keyword{datasets}
