% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{sel_items_src_rel}
\alias{sel_items_src_rel}
\title{Meta Data of Publication’s Sources from Max Planck Institutes}
\format{An object of class \code{"list"}}
\source{
\url{https://pure.mpg.de/rest}
}
\usage{
data(sel_items_src_rel)
}
\description{
A named list with nodes (persons, externals, ous) and edges between them and publication’s sources.
}
\keyword{datasets}
