% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{sel_stats}
\alias{sel_stats}
\title{Statistics of Publications from Max Planck Institutes}
\format{An object of class \code{"list"}}
\source{
\url{https://pure.mpg.de/rest}
}
\usage{
data(sel_stats)
}
\description{
A named list with different count data derived from the publication corpus:
records per institute, articles per journal and publications per person.
}
\keyword{datasets}
