% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{titles_eng}
\alias{titles_eng}
\title{Documents with English Publication Titles}
\format{An object of class \code{"VCorpus"}; see \code{\link[tm]{VCorpus}}.}
\source{
\url{https://pure.mpg.de/rest}
}
\usage{
data(titles_eng)
}
\description{
A plain text corpus consisting of more than 220.000 documents with one
publication title per document.
}
\keyword{datasets}
