% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{titles_mpi}
\alias{titles_mpi}
\title{Documents with English Publication Titles from Max Planck Institutes}
\format{An object of class \code{"VCorpus"}; see \code{\link[tm]{VCorpus}}.}
\source{
\url{https://pure.mpg.de/rest}
}
\usage{
data(titles_mpi)
}
\description{
A plain text corpus consisting of 73 documents with publication titles
from different Max Planck Institutes
}
\keyword{datasets}
