% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{titles_pers}
\alias{titles_pers}
\title{Documents with English Publication Titles from Scientists of Max Planck Institutes}
\format{An object of class \code{"VCorpus"}; see \code{\link[tm]{VCorpus}}.}
\source{
\url{https://pure.mpg.de/rest}
}
\usage{
data(titles_pers)
}
\description{
A plain text corpus consisting of more than 30.000 documents with publication titles
from different scientists of Max Planck Institutes
}
\keyword{datasets}
